requirements.txt
altgraph==0.17.2
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
atoti==0.6.5
atoti-core==0.6.5
atoti-jupyterlab==0.6.5
atoti-query==0.6.5
attrs==21.4.0
auto-py-to-exe==2.18.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Eel==0.12.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fpdf==1.7.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
jdk4py==11.0.13.1
jedi==0.18.1
Jinja2==3.1.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.3
jupyterlab-pygments==0.2.1
jupyterlab-server==2.12.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==6.0.1
pycparser==2.21
Pygments==2.11.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
stack-data==0.2.0
tenacity==8.0.1
terminado==0.13.3
tinycss2==1.1.1
tkcalendar==1.6.1
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
twilio==7.8.0
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
whichcraft==0.6.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
